pyproject.toml
[tool.poetry]
name = "img2bin"
version = "0.1.0"
description = ""
authors = ["okooo5km <okooo5km@gmail.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "default"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
typer = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
img2bin = "img2bin.main:main"
